Ajout de validation dans le calendarService

+++ Utils/DataServices/CalendarService.cs
@@ -7,15 +7,16 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using System.Xml.Linq;
 
 namespace Automate.Utils.Services
 {
     public interface ICalendarService
     {
-        public List<Models.TaskModel> GetTasksByDate(DateTime? date);
+        public List<TaskModel> GetTasksByDate(DateTime? date);
 
-        public void AddTask(Models.TaskModel task);
+        public void AddTask(TaskModel task);
         
         public void UpdateTask(string newTaskDescription, string taskId);
 
@@ -35,27 +36,89 @@ namespace Automate.Utils.Services
 
         public List<TaskModel> GetTasksByDate(DateTime? date)
         {
-            var filter = Builders<TaskModel>.Filter.Eq("Date", date);
-            var tasks = _tasks.Find(filter).ToList();
-            return tasks;
+            try
+            {
+                if (date is null)
+                    throw new ArgumentException("la date ne peut pas être null.");
+
+                var filter = Builders<TaskModel>.Filter.Eq("Date", date);
+                var tasks = _tasks.Find(filter).ToList();
+                return tasks;
+            }
+            catch (Exception e)
+            {
+                MessageBox.Show(e.Message,
+                   "Erreur",
+                   MessageBoxButton.OK,
+                   MessageBoxImage.Error);
+                return new List<TaskModel>();
+            }
         }
 
         public void AddTask(TaskModel task)
         {
-            _tasks.InsertOne(task);
+            try
+            {
+                if(task is null)
+                    throw new ArgumentException("La tâche ne peut pas être null");
+
+                if (task.Date is null)
+                    throw new ArgumentException("La date de la tâche ne doit pas être null.");
+
+                if (string.IsNullOrEmpty(task.Description))
+                    throw new ArgumentException("La description de la tâche ne doit pas être null.");
+
+                _tasks.InsertOne(task);
+            }
+            catch (Exception e)
+            {
+                MessageBox.Show(e.Message,
+                   "Erreur",
+                   MessageBoxButton.OK,
+                   MessageBoxImage.Error);
+            }
         }
 
         public void UpdateTask(string newTaskDescription, string taskId)
         {
-            var filter = Builders<TaskModel>.Filter.Eq("_id", taskId);
-            var update = Builders<TaskModel>.Update.Set("Description", newTaskDescription);
-            _tasks.UpdateOne(filter, update);
+            try
+            {
+                if (string.IsNullOrEmpty(taskId))
+                    throw new ArgumentException("L'identifiant de la tâche à modifier ne doit pas être vide");
+
+                if (string.IsNullOrEmpty(newTaskDescription))
+                    throw new ArgumentException("La nouvelle tâche ne doit pas être vide.");
+
+                var filter = Builders<TaskModel>.Filter.Eq("_id", taskId);
+                var update = Builders<TaskModel>.Update.Set("Description", newTaskDescription);
+                _tasks.UpdateOne(filter, update);
+            }
+            catch(Exception e)
+            {
+                MessageBox.Show(e.Message,
+                    "Erreur",
+                    MessageBoxButton.OK,
+                    MessageBoxImage.Error);
+            }
         }
 
         public void DeleteTask(string taskId)
         {
-            var filter = Builders<TaskModel>.Filter.Eq("_id", taskId);
-            _tasks.DeleteOne(filter);
+            try
+            {
+                if (string.IsNullOrEmpty(taskId))
+                    throw new ArgumentException("L'identifiant de la tâche à supprimer ne doit pas être vide");
+
+                var filter = Builders<TaskModel>.Filter.Eq("_id", taskId);
+                _tasks.DeleteOne(filter);
+            }
+            catch(Exception e)
+            {
+                MessageBox.Show(e.Message,
+                   "Erreur",
+                   MessageBoxButton.OK,
+                   MessageBoxImage.Error);
+            }
         }
     }
 }
